move general helpers to utils
