+++ packages/aws-client/utils.js
@@ -1,5 +1,67 @@
 const pRetry = require('p-retry');
+const { UnparsableFileLocationError } = require('@cumulus/common/errors');
 const { setErrorStack } = require('@cumulus/common/util');
+
+exports.findResourceArn = (obj, fn, prefix, baseName, opts, callback) => {
+  obj[fn](opts, (err, data) => {
+    if (err) {
+      callback(err, data);
+      return;
+    }
+
+    let arns = null;
+    Object.keys(data).forEach((prop) => {
+      if (prop.endsWith('Arns')) {
+        arns = data[prop];
+      }
+    });
+
+    if (!arns) {
+      callback(`Could not find an 'Arn' property in response from ${fn}`, data);
+      return;
+    }
+
+    const prefixRe = new RegExp(`^${prefix}-[A-Z0-9]`);
+    const baseNameOnly = `-${baseName}-`;
+    let matchingArn = null;
+
+    arns.forEach((arn) => {
+      const name = arn.split('/').pop();
+      if (name.match(prefixRe) && name.includes(baseNameOnly)) {
+        matchingArn = arn;
+      }
+    });
+
+    if (matchingArn) {
+      callback(null, matchingArn);
+    } else if (data.NextToken) {
+      const nextOpts = Object.assign({}, opts, { NextToken: data.NextToken });
+      exports.findResourceArn(obj, fn, prefix, baseName, nextOpts, callback);
+    } else {
+      callback(`Could not find resource ${baseName} in ${fn}`);
+    }
+  });
+};
+
+/**
+ * Extract the S3 bucket and key from the URL path parameters
+ *
+ * @param {string} pathParams - path parameters from the URL
+ * @returns {Object} - bucket/key in the form of
+ * { Bucket: x, Key: y }
+ */
+exports.getFileBucketAndKey = (pathParams) => {
+  const fields = pathParams.split('/');
+
+  const Bucket = fields.shift();
+  const Key = fields.join('/');
+
+  if (Bucket.length === 0 || Key.length === 0) {
+    throw new UnparsableFileLocationError(`File location "${pathParams}" could not be parsed`);
+  }
+
+  return [Bucket, Key];
+};
 
 /**
  * Wrap a function and provide a better stack trace
